+++ Engine/NativeHeaders.h
@@ -12,3 +12,5 @@
 extern int use_analog, joystick_move_sensitivity, joystick_turn_sensitivity;
 #define SDL_SwapBE16(x) OSSwapBigToHostInt16(x)
 #define SDL_SwapBE32(x) OSSwapBigToHostInt32(x)
+
+extern int MD_CampaignBossDeath(void *object);